+++ src/chrome/content/mboximport/menufunctions.js
@@ -100,6 +100,7 @@ function rfc822test() {
 
 
 function IETsetMBmenu() {
+    console.debug('IETsetMBmenu');
     document.getElementById("mboxexport").removeAttribute("disabled");
     var msgFolder = GetSelectedMsgFolders()[0];
     var isVirtFol = msgFolder ? msgFolder.flags & 0x0020 : false;
@@ -112,6 +113,7 @@ function IETsetMBmenu() {
 
     // the folder is the account pseudo-folder? we must set the right label and
     if (msgFolder.isServer) {
+        console.debug('isServer');
         document.getElementById("mboxexportallstruct").collapsed = false;
         document.getElementById("mboxexport").label = mboximportbundle.GetStringFromName("exportAccount");
         document.getElementById("mboxexportZIP").collapsed = true;
@@ -169,7 +171,7 @@ function IETsetMBmenu() {
 function IETsetMBmenu2(popup) {
     var i = popup.getAttribute("mboxIndex");
 
-    console.debug('no phones are selected bm2');
+    console.debug('IETsetMBmenu2');
     // if (!msgFolder) {
 	// 	alert(mboximportbundle.GetStringFromName("noFolderSelected"));
 	// }
@@ -211,6 +213,7 @@ function IETsetMBmenu2(popup) {
 
     // the folder is the account pseudo-folder? we must set the right label
     if (msgFolder.isServer) {
+        console.debug('isserver');
         document.getElementById("mboximportEML" + i).setAttribute("disabled", "true");
         document.getElementById("mboxexportsub" + i).collapsed = true;
         document.getElementById("mboxexportstruct" + i).collapsed = true;